rename conditional density utils, adapt imports.
